requirements.txt
Package             Version
------------------- -------
asgiref             3.3.1  
Django              3.1.5  
django-environ      0.4.5  
djangorestframework 3.12.2 
pip                 20.0.2 
pkg-resources       0.0.0  
psycopg2            2.8.6  
pytz                2020.5 
setuptools          44.0.0 
sqlparse            0.4.1  
wheel               0.36.2 
